+++ src/test/groovy/all/shared/gradle/quality/code/BaseStyleConfigWrapperTest.groovy
@@ -2,38 +2,26 @@
 //  Licensed under the MIT License (MIT), see LICENSE.txt
 package all.shared.gradle.quality.code
 
+import all.shared.gradle.testfixtures.SpyProjectFactory
+
 import groovy.transform.CompileStatic
 
 import org.gradle.api.Project
-import org.gradle.api.logging.Logger
-import org.gradle.testfixtures.ProjectBuilder
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.Dependency
 
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
 import static org.junit.jupiter.api.Assertions.assertEquals
 import static org.junit.jupiter.api.Assertions.assertTrue
 
 import static org.mockito.Matchers.eq
-import static org.mockito.Mockito.doReturn
-import static org.mockito.Mockito.mock
-import static org.mockito.Mockito.spy
 import static org.mockito.Mockito.verify
 
 @CompileStatic
 class BaseStyleConfigWrapperTest {
-  private final Project spyProject = spy(ProjectBuilder.builder().build())
-  private final Logger mockLogger = mock(Logger)
-
-  @BeforeEach
-  void beforeEachTest() {
-    doReturn(mockLogger)
-      .when(spyProject)
-      .getLogger()
-  }
+  private final Project spyProject = SpyProjectFactory.build()
 
   @Test
   void shouldAddExtension() {
@@ -42,12 +30,15 @@ class BaseStyleConfigWrapperTest {
     assertTrue(spyProject.properties['extensionName'] instanceof BaseStyleConfigWrapperExtension)
     assertEquals(
       BaseStyleConfigWrapper.QUALITY_CONFIG_MAVEN_REPO_URL,
-      ((MavenArtifactRepository) spyProject.repositories.getByName('maven')).getUrl().toString())
+      ((MavenArtifactRepository) spyProject.repositories.getByName('maven'))
+        .getUrl()
+        .toString())
     final Configuration configuration = spyProject.configurations.getByName('extensionName')
-    assertTrue(configuration.allDependencies
-      .collect { Dependency dependency -> "$dependency.group:$dependency.name:$dependency.version" }
-      .contains("$BaseStyleConfigWrapper.QUALITY_CONFIG_MAVEN_COORDINATES:+"))
-    verify(mockLogger).debug(eq('Added base-style-config-wrapper extension'))
+    assertTrue(configuration.allDependencies.stream()
+      .anyMatch { final Dependency dependency ->
+        "$dependency.group:$dependency.name:$dependency.version" == "$BaseStyleConfigWrapper.QUALITY_CONFIG_MAVEN_COORDINATES:+" })
+    verify(spyProject.logger)
+      .debug(eq('Added base-style-config-wrapper extension'))
   }
 
   @Test
@@ -57,9 +48,9 @@ class BaseStyleConfigWrapperTest {
     BaseStyleConfigWrapper.addExtension(spyProject, 'extensionName')
 
     final Configuration configuration = spyProject.configurations.getByName('extensionName')
-    assertTrue(configuration.allDependencies
-      .collect { Dependency dependency -> "$dependency.group:$dependency.name:$dependency.version" }
-      .contains("$BaseStyleConfigWrapper.QUALITY_CONFIG_MAVEN_COORDINATES:1.0.4"))
+    assertTrue(configuration.allDependencies.stream()
+      .anyMatch { final Dependency dependency ->
+        "$dependency.group:$dependency.name:$dependency.version" == "$BaseStyleConfigWrapper.QUALITY_CONFIG_MAVEN_COORDINATES:1.0.4" })
   }
 
   @Test
@@ -68,6 +59,7 @@ class BaseStyleConfigWrapperTest {
 
     BaseStyleConfigWrapper.addExtension(spyProject, 'extensionName')
 
-    verify(mockLogger).error(eq('Couldn\'t add base-style-config-wrapper extension'))
+    verify(spyProject.logger)
+      .error(eq('Couldn\'t add base-style-config-wrapper extension'))
   }
 }