+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Base Style Config Wrapper Change Log
 
+## 1.0.2 - March 2019
+
+* Updates main code to work with base-style-config:2.0.0.
+* Updates test code to use [spy-project-factory](https://github.com/gmullerb/spy-project-factory).
+* Removes redundant getter code.
+* Adds Gitlab CI.
+* Adds Versions Compatibility Table.
+* Upgrades Gradle to version 5.3.
+* Updates README file.
 
 ## 1.0.1 - October 2018